oof....
after delegating to an async init from an async convenience init,
we are now having SILGen emit a hop _right_ after that apply, because it's
in the executor breadcrumb. the problem is that we haven't even assigned
self to the box yet, in the value constructor! so, the load is wrong because the box
has nothing in it yet; even if the load would verify in SIL.  so, we really can't rely on
the executor breadcrumb for general async functions, unless if we do some
more refactoring.