introduce `@_layout` protocols
A layout protocol is like a marker protocol, but with
a number of unique characteristics that make it different:

- A layout protocol has some form of runtime metadata
to facilitate dynamic casts and type-tests. This runtime
metadata is not necessarily a standard protocol conformance
descriptor, either.

- A conformance to a layout protocol can be automatically
derived in a more aggressive way than a standard marker
protocol, because a "layout" is an implementation detail.

Keep in mind that these layout protocols are very much
like a LayoutConstraint that is fully-integrated with
the type system.