[SIL] fix existential cast optimization
The new instruction created by this optimization has an implicit
assumption that the number of conformances required for an
`init_existential_addr` is equal to 1 if the constraint type of the
existential is a single protocol. That's not quite accurate, as the
protocols listed in the ExistentialLayout is what really describes what
conformances are required. While SILGen and the SILVerifier enforced
that already, this cast optimizer was missing conformances to Copyable
and Escapable that are only listed in the ExistentialLayout.